requirements.txt
scipy
scikit-image
matplotlib
pillow
numpy